dom/canvas/test/test_offscreencanvas_neuter.html

<!DOCTYPE HTML>
<html>
<head>
<title>OffscreenCanvas: Test neutering</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<canvas id="c" width="64" height="64"></canvas>
<script>

SimpleTest.waitForExplicitFinish();

function runTest() {

  var htmlCanvas = document.getElementById("c");
  var worker = new Worker("offscreencanvas_neuter.js");

  ok(htmlCanvas, "Should have HTML canvas element");
  ok(worker, "Web worker successfully created");

  ok(htmlCanvas.transferControlToOffscreen, "HTMLCanvasElement has transferControlToOffscreen function");

  var offscreenCanvas = htmlCanvas.transferControlToOffscreen();
  ok(offscreenCanvas, "Expected transferControlToOffscreen to succeed");

  /* check html canvas is neuterd */
  is(htmlCanvas.width, 64, "HTML canvas has correct width");
  SimpleTest.doesThrow(
    function() { htmlCanvas.width = 128; },
    "Can't change html canvas' width after transferControlToOffscreen");

  SimpleTest.doesThrow(
    function() { htmlCanvas.height = 128; },
    "Can't change html canvas' height after transferControlToOffscreen");

  ok(!htmlCanvas.getContext("2d"), "Can't getContext after transferControlToOffscreen");
  ok(!htmlCanvas.getContext("webgl"), "Can't getContext after transferControlToOffscreen");
  ok(!htmlCanvas.getContext("webgl2"), "Can't getContext after transferControlToOffscreen");

  worker.postMessage(offscreenCanvas, [offscreenCanvas]);

  /* check parent offscreencanvas is neutered after being transfered */
  SimpleTest.doesThrow(
    function() { offscreenCanvas.width = 128; },
    "Can't change transfered worker canvas width");

  SimpleTest.doesThrow(
    function() { offscreenCanvas.height = 128; },
    "Can't change transfered worker canvas height");

  SimpleTest.doesThrow(
    function() { offscreenCanvas.getContext("2d") },
    "Can't getContext on transfered worker canvas");

  SimpleTest.doesThrow(
    function() { offscreenCanvas.getContext("webgl") },
    "Can't getContext on transfered worker canvas");

  SimpleTest.doesThrow(
    function() { offscreenCanvas.getContext("webgl2") },
    "Can't getContext on transfered worker canvas");

  // Transfer a neutered offscreencanvas should be ok.
  worker.postMessage(offscreenCanvas, [offscreenCanvas]);

  worker.terminate();
  SimpleTest.finish();
}

SpecialPowers.pushPrefEnv({'set': [
  ['gfx.offscreencanvas.enabled', true],
  ['webgl.force-enabled', true],
]}, runTest);

</script>
</body>
</html>
